fix(technologies-2): fix tooltips in TechnologyCategorySubComponent and changes Types `int` to `double` in `AggregateItemDto` and `Resource`

+++ src/main/java/de/uniks/stp24/model/Resource.java
@@ -3,7 +3,7 @@ package de.uniks.stp24.model;
 
 public record Resource(
     String resourceID,
-    int count,
-    int changePerSeason
+    double count,
+    double changePerSeason
 ) {
 }
